+++ CHANGELOG.md
@@ -18,11 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - make SPI `write` method public
 - `set_temp_monitor` method
 - `set_lna_gain` method
-- `set_ocp`, `set_ocp_trim` methods
+- `set_ocp` method
 - `set_crc` method and make `set_coding_rate` public
 - `rssi`, `last_packet_rssi` methods
 - `set_low_data_rate_optimize` method
 - `set_preamble_length` method
+- `set_pa_boost`, `set_pa_rfo` methods
 
 ## [0.1.0] - 2026-03-07